adding card collection interface and editing PlayerHand class

+++ src/main/models/Card.java
@@ -2,7 +2,7 @@ package main.models;
 
 
 public class Card {
-    private final double value;
+    private final int value;
     private final String color;
     private boolean isFaceUp;
